Server: Bug fixes.

+++ server/database/migrations/2023_11_08_143590_create_user_roles_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('user_roles', function (Blueprint $table) {
-            $table->integer('id')->primary();
+            $table->unsignedBigInteger('id')->primary();
             $table->string('title');
         });
     }
